Add command to perform local extraction
A local extraction processes all the localizable assets, extracts their text units and save them into a JSON file. The JSON file name is the asset filename with ".json" appended at the end. The local extraction is given a name (-n option). That name is used to created a directory (inside the output directory) that will hold all the JSON files. This way it is possible to have multiple extractions for the same code base and then make diff between them to see if text unit were added/removed, etc.

The JSON format looks like:

 {
   "localTextUnits" : [ {
     "name" : "100 character description: --- 100_character_description_",
     "source" : "100 character description:",
     "comments" : null
   },
   ...
   {
     "name" : "There is {number} car --- car _other",
     "source" : "There are {number} cars",
     "comments" : "Test plural"
   } ],
   "filterOptions" : [ "testoption=something" ],
   "fileType" : "POFileType",
   "name" : "source1"
 }

In some way the command is similar to the push command but instead of calling the server to process the assets and create/save text units it does everything locally and persist the text unit locally in a JSON file. This will allow to identify change locally when processing pull requests hence without putting any load on the server or depend on its uptime.